Change the branch 1.0.2 to 1.0.3.

+++ components/dfs/filesystems/uffs/src/inc/uffs/uffs_blockinfo.h
@@ -51,7 +51,7 @@ extern "C"{
  */
 struct uffs_PageSpareSt {
 	uffs_Tags tag;			//!< page tag
-	int expired:1;
+	u8 expired:1;
 };
 
 /**
